Data Analysis Project- Excel Dashboard.
In this project, we will use a dataset from ‘Call Center Data’ located here. This time, I took the same dataset and build a dashboard for it using Excel. Reason for use Excel? hmmm, Excel is still a widely used tool for data analysis and it’s always a good tool to have under your belt. So why not? plus it’s fun. I mean it.